refactor: Applied several changes
- Moved TimerLoadingPage to /src/TimerPage from /src/TimerPage/component
- Set css file to print CR/LF correctly
- Changed 'useQuery' to 'useGetParliamentaryTableData' hook
