+++ src/index.css
@@ -7,6 +7,11 @@
     @apply font-pretendard;
   }
 }
+
+p {
+  white-space: pre-wrap;
+}
+
 .gradient-timer-running {
   background-image: linear-gradient(270deg, #21c494, #21bdc4, #21c45d);
 }
